Account: extract `GetServiceToken` flow helper, split out and rename its Mojom
- extract service-token logic off `LoggedInState` into a `GetServiceToken` flow helper; owned by `LoggedInState`
- split the service-token types out of `brave_account.mojom` into a dedicated `get_service_token.mojom`
- move the service-token unit tests to `flows/get_service_token_unittest.cc`
- register the new `get_service_token.mojom` bindings with all consuming bundles (brave://account, and Settings on all platforms), bumping the brave_account/resources GRD id allocation to fit the extra generated resource
