+++ browser/resources/settings/sources.gni
@@ -225,6 +225,7 @@ brave_settings_mojo_files = [
   "$root_gen_dir/brave/components/brave_account/mojom/brave_account.mojom-webui.ts",
   "$root_gen_dir/brave/components/brave_account/mojom/brave_account_row.mojom-webui.ts",
   "$root_gen_dir/brave/components/brave_account/mojom/change_password.mojom-webui.ts",
+  "$root_gen_dir/brave/components/brave_account/mojom/get_service_token.mojom-webui.ts",
   "$root_gen_dir/brave/components/brave_account/mojom/login.mojom-webui.ts",
   "$root_gen_dir/brave/components/brave_account/mojom/register.mojom-webui.ts",
   "$root_gen_dir/brave/components/brave_account/mojom/resend_verification_email.mojom-webui.ts",
