+++ components/brave_account/BUILD.gn
@@ -17,6 +17,8 @@ source_set("brave_account") {
     "brave_account_utils.h",
     "flows/change_password.cc",
     "flows/change_password.h",
+    "flows/get_service_token.cc",
+    "flows/get_service_token.h",
     "flows/login.cc",
     "flows/login.h",
     "flows/register.cc",
@@ -127,6 +129,7 @@ source_set("unit_tests") {
     "brave_account_service_test.h",
     "brave_account_service_unittest.cc",
     "flows/change_password_unittest.cc",
+    "flows/get_service_token_unittest.cc",
     "flows/login_unittest.cc",
     "flows/register_unittest.cc",
     "flows/resend_verification_email_unittest.cc",
