+++ components/brave_account/logged_in_state.cc
@@ -7,13 +7,9 @@
 
 #include <utility>
 
-#include "base/check.h"
-#include "base/check_deref.h"
 #include "base/functional/bind.h"
 #include "base/location.h"
-#include "base/types/expected.h"
 #include "brave/components/brave_account/brave_account_service_constants.h"
-#include "brave/components/brave_account/brave_account_utils.h"
 #include "brave/components/brave_account/endpoint_client/with_headers.h"
 #include "brave/components/brave_account/endpoints/auth_logout.h"
 #include "brave/components/brave_account/state_internal.h"
@@ -26,10 +22,7 @@ using endpoint_client::SetBearerToken;
 using endpoint_client::WithHeaders;
 using endpoints::AuthLogout;
 using endpoints::AuthValidate;
-using endpoints::ServiceToken;
-using internal::MakeClientError;
 using internal::MakeRequest;
-using internal::MakeServerError;
 
 LoggedInState::LoggedInState(
     AccountStatePrefs& account_state_prefs,
@@ -85,81 +78,7 @@ void LoggedInState::LogOut() {
 
 void LoggedInState::GetServiceToken(mojom::Service service,
                                     GetServiceTokenCallback callback) {
-  CHECK(service != mojom::Service::kAccounts);
-  std::string service_name(kServiceToString.at(service));
-  if (auto service_token =
-          Decrypt(account_state_prefs_->GetCachedServiceToken(service_name));
-      !service_token.empty()) {
-    return std::move(callback).Run(
-        mojom::GetServiceTokenResult::New(std::move(service_token)));
-  }
-
-  auto authentication_token =
-      GetDecryptedAuthenticationToken<mojom::GetServiceTokenError>();
-  if (!authentication_token.has_value()) {
-    return std::move(callback).Run(
-        base::unexpected(std::move(authentication_token).error()));
-  }
-
-  auto request = MakeRequest<WithHeaders<ServiceToken::Request>>();
-  SetBearerToken(request, *authentication_token);
-  request.body.service = service_name;
-
-  SendStateOwnedRequest<ServiceToken>(
-      std::move(request),
-      base::BindOnce(&LoggedInState::OnGetServiceToken,
-                     weak_factory_.GetWeakPtr(), std::move(service_name),
-                     std::move(callback)));
-}
-
-void LoggedInState::OnGetServiceToken(const std::string& service_name,
-                                      GetServiceTokenCallback callback,
-                                      ServiceToken::Response response) {
-  if (!response.body) {
-    return std::move(callback).Run(
-        base::unexpected(MakeServerError<mojom::GetServiceTokenError>(
-            response.status_code.value_or(response.net_error),
-            mojom::GetServiceTokenServerErrorCode::kInvalidResponse)));
-  }
-
-  const auto status_code = CHECK_DEREF(response.status_code);
-
-  auto result =
-      std::move(*response.body)
-          // expected<SuccessBody, [ErrorBody              ]> ==>
-          // expected<SuccessBody, [GetServiceTokenErrorPtr]>
-          .transform_error([&](auto error_body) {
-            return MakeServerError<mojom::GetServiceTokenError>(
-                status_code, std::move(error_body));
-          })
-          // expected<[SuccessBody             ], GetServiceTokenErrorPtr> ==>
-          // expected<[GetServiceTokenResultPtr], GetServiceTokenErrorPtr>
-          .and_then([&](auto success_body)
-                        -> base::expected<mojom::GetServiceTokenResultPtr,
-                                          mojom::GetServiceTokenErrorPtr> {
-            if (success_body.auth_token.empty()) {
-              return base::unexpected(
-                  MakeServerError<mojom::GetServiceTokenError>(
-                      status_code,
-                      mojom::GetServiceTokenServerErrorCode::kInvalidResponse));
-            }
-
-            auto encrypted_service_token = Encrypt(success_body.auth_token);
-            if (encrypted_service_token.empty()) {
-              return base::unexpected(
-                  MakeClientError<mojom::GetServiceTokenError>(
-                      mojom::GetServiceTokenClientErrorCode::
-                          kServiceTokenEncryptionFailed));
-            }
-
-            account_state_prefs_->CacheServiceToken(
-                service_name, std::move(encrypted_service_token));
-
-            return mojom::GetServiceTokenResult::New(
-                std::move(success_body.auth_token));
-          });
-
-  std::move(callback).Run(std::move(result));
+  get_service_token_(service, std::move(callback));
 }
 
 void LoggedInState::ScheduleAuthValidate(
