+++ components/brave_account/resources/BUILD.gn
@@ -70,6 +70,7 @@ build_webui("build") {
   mojo_files = [
     "$root_gen_dir/brave/components/brave_account/mojom/brave_account.mojom-webui.ts",
     "$root_gen_dir/brave/components/brave_account/mojom/change_password.mojom-webui.ts",
+    "$root_gen_dir/brave/components/brave_account/mojom/get_service_token.mojom-webui.ts",
     "$root_gen_dir/brave/components/brave_account/mojom/login.mojom-webui.ts",
     "$root_gen_dir/brave/components/brave_account/mojom/register.mojom-webui.ts",
     "$root_gen_dir/brave/components/brave_account/mojom/resend_verification_email.mojom-webui.ts",
