+++ ios/BUILD.gn
@@ -130,6 +130,7 @@ brave_core_public_headers = [
   "//ui/base/l10n/l10n_util_mac_bridge.h",
   "$root_gen_dir/brave/components/brave_account/mojom/brave_account.mojom.objc.h",
   "$root_gen_dir/brave/components/brave_account/mojom/change_password.mojom.objc.h",
+  "$root_gen_dir/brave/components/brave_account/mojom/get_service_token.mojom.objc.h",
   "$root_gen_dir/brave/components/brave_account/mojom/login.mojom.objc.h",
   "$root_gen_dir/brave/components/brave_account/mojom/register.mojom.objc.h",
   "$root_gen_dir/brave/components/brave_account/mojom/resend_verification_email.mojom.objc.h",
