+++ resources/resource_ids.spec
@@ -223,7 +223,7 @@
     "includes": [53980],
   },
   "<(SHARED_INTERMEDIATE_DIR)/brave/components/brave_account/resources/resources.grd": {
-    "META": {"sizes": {"includes": [45]}},
+    "META": {"sizes": {"includes": [50]}},
     "includes": [54000],
   },
   "brave/ios/web/test/test_resources.grd": {
